Major refactor: split alt src into multiple files
- Split up large alt.js file into multiple files, grouped by
  functionality. This should help make the code base easier to read and
  make changes to in the future (and hopefully help make it less
  intimidating for contributors)
