+++ src/action/ActionCreator.js
@@ -1,0 +1,16 @@
+'use strict'
+
+import { ACTION_HANDLER, ACTION_UID} from '../../src/shared/symbols'
+
+export default class ActionCreator {
+  constructor(alt, name, action, actions) {
+    this[ACTION_UID] = name
+    this[ACTION_HANDLER] = action.bind(this)
+    this.actions = actions
+    this.alt = alt
+  }
+
+  dispatch(data) {
+    this.alt.dispatch(this[ACTION_UID], data)
+  }
+}
